LPS-183358 Moves to manager package
